backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
psycopg2-binary
python-dotenv
httpx
